+++ service/controller/legacy/controller.go
@@ -31,7 +31,6 @@ type ControllerConfig struct {
 	Provider                string
 	Region                  string
 	Registry                string
-	IngressAPIVersion       string
 
 	OpsgenieKey string