+++ service/controller/managementcluster/resource.go
@@ -18,13 +18,11 @@ import (
 	"github.com/giantswarm/prometheus-meta-operator/v2/service/controller/resource/alerting/heartbeat"
 	"github.com/giantswarm/prometheus-meta-operator/v2/service/controller/resource/alerting/heartbeatwebhookconfig"
 	etcdcertificates "github.com/giantswarm/prometheus-meta-operator/v2/service/controller/resource/etcd-certificates"
-	ingressv1 "github.com/giantswarm/prometheus-meta-operator/v2/service/controller/resource/monitoring/ingress/v1"
-	ingressv1beta1 "github.com/giantswarm/prometheus-meta-operator/v2/service/controller/resource/monitoring/ingress/v1beta1"
+	"github.com/giantswarm/prometheus-meta-operator/v2/service/controller/resource/monitoring/ingress"
 	"github.com/giantswarm/prometheus-meta-operator/v2/service/controller/resource/monitoring/prometheus"
 	"github.com/giantswarm/prometheus-meta-operator/v2/service/controller/resource/monitoring/pvcresizingresource"
 	"github.com/giantswarm/prometheus-meta-operator/v2/service/controller/resource/monitoring/remotewriteapiendpointconfigsecret"
-	remotewriteingressv1 "github.com/giantswarm/prometheus-meta-operator/v2/service/controller/resource/monitoring/remotewriteingress/v1"
-	remotewriteingressv1beta1 "github.com/giantswarm/prometheus-meta-operator/v2/service/controller/resource/monitoring/remotewriteingress/v1beta1"
+	"github.com/giantswarm/prometheus-meta-operator/v2/service/controller/resource/monitoring/remotewriteingress"
 	"github.com/giantswarm/prometheus-meta-operator/v2/service/controller/resource/monitoring/remotewriteingressauth"
 	"github.com/giantswarm/prometheus-meta-operator/v2/service/controller/resource/monitoring/scrapeconfigs"
 	"github.com/giantswarm/prometheus-meta-operator/v2/service/controller/resource/monitoring/verticalpodautoscaler"
@@ -50,7 +48,6 @@ type resourcesConfig struct {
 	Provider                string
 	Region                  string
 	Registry                string
-	IngressAPIVersion       string
 
 	AlertmanagerAddress     string
 	AlertmanagerBaseDomain  string
@@ -248,8 +245,8 @@ func newResources(config resourcesConfig) ([]resource.Interface, error) {
 	}
 
 	var monitoringIngressResource resource.Interface
-	if config.IngressAPIVersion == "networking.k8s.io/v1beta1" {
-		c := ingressv1beta1.Config{
+	{
+		c := ingress.Config{
 			K8sClient:               config.K8sClient,
 			Logger:                  config.Logger,
 			BaseDomain:              config.PrometheusBaseDomain,
@@ -257,20 +254,7 @@ func newResources(config resourcesConfig) ([]resource.Interface, error) {
 			WhitelistedSubnets:      config.WhitelistedSubnets,
 		}
 
-		monitoringIngressResource, err = ingressv1beta1.New(c)
-		if err != nil {
-			return nil, microerror.Mask(err)
-		}
-	} else {
-		c := ingressv1.Config{
-			K8sClient:               config.K8sClient,
-			Logger:                  config.Logger,
-			BaseDomain:              config.PrometheusBaseDomain,
-			RestrictedAccessEnabled: config.RestrictedAccessEnabled,
-			WhitelistedSubnets:      config.WhitelistedSubnets,
-		}
-
-		monitoringIngressResource, err = ingressv1.New(c)
+		monitoringIngressResource, err = ingress.New(c)
 		if err != nil {
 			return nil, microerror.Mask(err)
 		}
@@ -320,25 +304,14 @@ func newResources(config resourcesConfig) ([]resource.Interface, error) {
 	}
 
 	var remoteWriteIngressResource resource.Interface
-	if config.IngressAPIVersion == "networking.k8s.io/v1beta1" {
-		c := remotewriteingressv1beta1.Config{
+	{
+		c := remotewriteingress.Config{
 			K8sClient:  config.K8sClient,
 			Logger:     config.Logger,
 			BaseDomain: config.PrometheusBaseDomain,
 		}
 
-		remoteWriteIngressResource, err = remotewriteingressv1beta1.New(c)
-		if err != nil {
-			return nil, microerror.Mask(err)
-		}
-	} else {
-		c := remotewriteingressv1.Config{
-			K8sClient:  config.K8sClient,
-			Logger:     config.Logger,
-			BaseDomain: config.PrometheusBaseDomain,
-		}
-
-		remoteWriteIngressResource, err = remotewriteingressv1.New(c)
+		remoteWriteIngressResource, err = remotewriteingress.New(c)
 		if err != nil {
 			return nil, microerror.Mask(err)
 		}