+++ CHANGELOG.md
@@ -7,9 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Do not reconcile service monitors in CAPO MCs.
+
 ### Removed
 
 - Remove option to disable PVC creations only used on KVM installations.
+- Remove deprecated ingress v1beta1 (only used on kvm).
 
 ## [4.12.0] - 2022-11-25